organized code more, fixed imports, and seperated code into proper containers
