+++ client/components/Navbar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
-export default function Navbar() {
+export function Navbar() {
   const navigate = useNavigate();
   function home() {
     navigate('/Home');
