+++ client/components/Searchbar.jsx
@@ -1,3 +1,3 @@
-export default function searchBar() {
+export function searchBar() {
   return <input type="text" placeholder="Submit your search here..."></input>;
 }
