+++ client/pages/Comments.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { Editor } from '@tinymce/tinymce-react';
 import { useParams } from 'react-router-dom';
+import { Navbar } from '../components/Navbar.jsx';
 import '../styles/Comments.scss';
-import Navbar from '../components/Navbar.jsx';
 import HelperFunctions from '../helper-functions';
 
 const Comments = () => {
